requirements.txt
torch
torchvision
torchnet
numpy
Pillow
tqdm